Update number of frame required for processing
The number of frames required for processing was based on testing with
MATLAB 2021b. In that case, without bounding time frames, there were
small floating point differences in the results. In MATLAB 2024b, that
issue has been resolved and it is possible to process a single time
point for spatial derivates and structure tensor inputs.

For temporal derivates, the initial requirement was Nt > 3*tSig.
However, this is misisng a factor of 2, and it should be > 6*tSig. This
is because the corresponding filter is of radius 3*tSig, so 6*tSig+1
total in size. To avoid edge effects, this many frames must be processed
 together.

This update also splits the dtI calculation into two steps to better
reduce memory usage.