Calculate 3D eigenvalues numerically
Previously the eigenvalues used to calculate reliability were calculated
by analytically solving for the eigenvalues of the 3x3 matrix. This was
because in the original, MATLAB 2021 version, pageeig did not exist as
a function. This implementation had a few issues, including that
there appear to have been typos in the equations for L2 and L3. Beyond
these typos, this analytical solution let to differences in the MATLAB
and Python reliabilities, perhaps because of the propagation of
uncertanties in the difference between the the two implementations of
image filtering.

With this commit, both MATLAB and Python use built in functions for
numerical calculation of the eigenvalues (pageeig and np.linalg.eigvals,
respectively).