Use TIFFvolume for all image loading
Previously imread_big was used to load in ImageJ files that consist of
one large tif. This was leading to some errors in loading some pixels.
That seems to be caused by the +1 on the calculation of the starting
points.

In the new version, TIFFvolume uses the TIFF library to open a volume
unless it detect that there is ImageJ metadata and that TIFF thinks the
file only has one frame. In that case, fread is used to manually load
the frames. This is based on assumptions about the ImageJ format, and
may or may not be generalizable.